// src/core/types/chaiBuilderEditorProps.ts
export type SavePageData = {
  autoSave: boolean;
  blocks: ChaiBlock[];
  theme?: ChaiBuilderThemeValues;
};
// ...
export type ChaiBuilderThemeValues = {
  fontFamily: Record<string, string>;
  borderRadius: string;
  colors: Record<string, string[]>;
};